More doc comments.
