+++ server/swimos_connector_kafka/src/config.rs
@@ -26,26 +26,46 @@ use crate::{
     error::DerserializerLoadError,
 };
 
+/// Configuration parameters for the Kafka connector.
 #[derive(Clone, Debug, Form)]
 #[form(tag = "kafka")]
 pub struct KafkaConnectorConfiguration {
+    /// Properties to configure the Kafka consumer.
     pub properties: HashMap<String, String>,
+    /// Log level for the Kafka consumer.
     pub log_level: KafkaLogLevel,
+    /// Specifications for the value lanes to define for the connector. This includes a pattern to define a selector
+    /// that will pick out values to set to that lane, from a Kafka message.
     pub value_lanes: Vec<ValueLaneSpec>,
+    /// Specifications for the map lanes to define for the connector. This includes a pattern to define a selector
+    /// that will pick out updates to apply to that lane, from a Kafka message.
     pub map_lanes: Vec<MapLaneSpec>,
+    /// Deserialization format to use to interpret the contents of the keys of the Kafka messages.
     pub key_deserializer: DeserializationFormat,
+    /// Deserialization format to use to interpret the contents of the payloads of the Kafka messages.
     pub payload_deserializer: DeserializationFormat,
+    /// A list of Kafka topics to subscribe to.
     pub topics: Vec<String>,
 }
 
+/// Specification of a value lane for the Kafka connector.
 #[derive(Clone, Debug, Form)]
 pub struct ValueLaneSpec {
+    /// A name to use for the lane. If not specified, the connector will attempt to infer one from the selector.
     pub name: Option<String>,
+    /// String representation of a selector to extract values for the lane from Kafka messages.
     pub selector: String,
+    /// Whether the lane is required. If this is `true` and the selector returns nothing for a Kafka Message, the
+    /// connector will fail with an error.
     pub required: bool,
 }
 
 impl ValueLaneSpec {
+    /// # Arguments
+    /// * `name` - A name to use for the lane. If not specified the connector will attempt to infer a name from the selector.
+    /// * `selector` - String representation of the selector to extract values from the Kafka message.
+    /// * `required` - Whether the lane is required. If this is `true` and the selector returns nothing for a Kafka Message, the
+    /// connector will fail with an error.
     pub fn new<S: Into<String>>(name: Option<S>, selector: S, required: bool) -> Self {
         ValueLaneSpec {
             name: name.map(Into::into),
@@ -55,16 +75,32 @@ impl ValueLaneSpec {
     }
 }
 
+/// Specification of a value lane for the Kafka connector.
 #[derive(Clone, Debug, Form)]
 pub struct MapLaneSpec {
+    /// The name of the lane.
     pub name: String,
+    /// String representation of a selector to extract the map keys from the Kafka messages.
     pub key_selector: String,
+    /// String representation of a selector to extract the map values from the Kafka messages.
     pub value_selector: String,
+    /// Whether to remove an entry from the map if the value selector does not return a value. Otherwise, missing
+    /// values will be treated as a failed extraction from the message.
     pub remove_when_no_value: bool,
+    /// Whether the lane is required. If this is `true` and the selector returns nothing for a Kafka Message, the
+    /// connector will fail with an error.
     pub required: bool,
 }
 
 impl MapLaneSpec {
+    /// # Arguments
+    /// * `name` - The name of the lane.
+    /// * `key_selector` - String representation of a selector to extract the map keys from the Kafka messages.
+    /// * `value_selector` - String representation of a selector to extract the map values from the Kafka messages.
+    /// * `remove_when_no_value` - Whether to remove an entry from the map if the value selector does not return a value. Otherwise, missing
+    /// values will be treated as a failed extraction from the message.
+    /// * `required` - Whether the lane is required. If this is `true` and the selector returns nothing for a Kafka Message, the
+    /// connector will fail with an error.
     pub fn new<S: Into<String>>(
         name: S,
         key_selector: S,
@@ -82,6 +118,7 @@ impl MapLaneSpec {
     }
 }
 
+/// Supported deserialization formats to use to interpret a component of a Kafka message.
 #[derive(Clone, Form, Debug, Default)]
 pub enum DeserializationFormat {
     #[default]
@@ -99,11 +136,13 @@ pub enum DeserializationFormat {
     Json,
     #[cfg(feature = "avro")]
     Avro {
+        /// If this is specified, loading the deserializer will attempt to load an Avro schema from a file at this path.
         schema_path: Option<String>,
     },
 }
 
 impl DeserializationFormat {
+    /// Attempt to load a deserializer based on the format descriptor.
     pub async fn load(&self) -> Result<BoxMessageDeserializer, DerserializerLoadError> {
         match self {
             DeserializationFormat::Bytes => Ok(BytesDeserializer.boxed()),
@@ -148,6 +187,7 @@ impl DeserializationFormat {
     }
 }
 
+/// Enumeration of logging levels supported by the underlying Kafka consumer.
 #[derive(Clone, Copy, Debug, Form, PartialEq, Eq)]
 pub enum KafkaLogLevel {
     Emerg,
