+++ server/swimos_connector_kafka/src/connector/tests/mod.rs
@@ -12,6 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+#[cfg(feature = "json")]
 mod end_to_end;
 mod integration;
