+++ server/swimos_connector_kafka/src/lib.rs
@@ -22,14 +22,9 @@ mod selector;
 pub use config::{
     DeserializationFormat, KafkaConnectorConfiguration, KafkaLogLevel, MapLaneSpec, ValueLaneSpec,
 };
-pub use connector::{KafkaConnector, MessagePart};
-#[cfg(feature = "avro")]
-pub use deser::AvroDeserializer;
-#[cfg(feature = "json")]
-pub use deser::JsonDeserializer;
-pub use deser::{
-    BoxErrorDeserializer, BoxMessageDeserializer, BytesDeserializer, Endianness, F32Deserializer,
-    F64Deserializer, I32Deserializer, I64Deserializer, MessageDeserializer, MessageView,
-    ReconDeserializer, StringDeserializer, U32Deserializer, U64Deserializer, UuidDeserializer,
+pub use connector::KafkaConnector;
+pub use deser::Endianness;
+pub use error::{
+    BadSelector, DerserializerLoadError, DeserializationError, InvalidLaneSpec, InvalidLanes,
+    KafkaConnectorError, LaneSelectorError,
 };
-pub use error::KafkaConnectorError;
